travel_ml/requirements.txt
tqdm~=4.66.2
pandas
numpy
fastapi~=0.110.0
uvicorn
clickhouse-connect
transformers~=4.38.2
stability_sdk
boto3~=1.34.69
diffusers~=0.27.2
accelerate
sentencepiece
Cython
redis
loguru~=0.7.2
ruclip==0.0.2
rembg[gpu]
torch~=1.13.1
torchvision==0.14.1+cu116
torchaudio==0.13.1 
pydantic~=2.6.4
starlette~=0.36.3
python-dotenv~=1.0.1
pillow~=10.2.0
requests~=2.31.0